moved add_document_handler and tensor_fields_container around
