+++ src/marqo/core/index_management/index_management.py
@@ -18,7 +18,7 @@ from marqo.core.models import MarqoIndex
 from marqo.core.models.marqo_index import SemiStructuredMarqoIndex
 from marqo.core.models.marqo_index_request import MarqoIndexRequest
 from marqo.core.semi_structured_vespa_index.semi_structured_vespa_schema import SemiStructuredVespaSchema
-from marqo.core.vespa_schema import for_marqo_index_request as vespa_schema_factory
+from marqo.core.vespa_index.vespa_schema import for_marqo_index_request as vespa_schema_factory
 from marqo.tensor_search.models.index_settings import IndexSettings
 from marqo.vespa.exceptions import VespaStatusError
 from marqo.vespa.vespa_client import VespaClient
