+++ src/marqo/core/semi_structured_vespa_index/semi_structured_vespa_schema.py
@@ -4,7 +4,7 @@ from jinja2 import Environment, FileSystemLoader
 
 from marqo.core.models.marqo_index import SemiStructuredMarqoIndex, MarqoIndex
 from marqo.core.models.marqo_index_request import UnstructuredMarqoIndexRequest
-from marqo.core.vespa_schema import VespaSchema
+from marqo.core.vespa_index.vespa_schema import VespaSchema
 
 
 class SemiStructuredVespaSchema(VespaSchema):
