+++ tests/core/inference/test_tensor_fields_container.py
@@ -4,7 +4,7 @@ from typing import cast
 import pytest
 
 from marqo.core.constants import MARQO_DOC_ID, MARQO_DOC_TENSORS, MARQO_DOC_CHUNKS, MARQO_DOC_EMBEDDINGS
-from marqo.core.document.tensor_fields_container import TensorFieldsContainer, MultiModalTensorFieldContent
+from marqo.core.inference.tensor_fields_container import TensorFieldsContainer, MultiModalTensorFieldContent
 from marqo.core.exceptions import AddDocumentsError
 from marqo.core.models.marqo_index import FieldType
